refactor(query-performance-ai): rename _runtime.py to runtime.py
Drop the leading underscore: the rest of products/query_performance_ai/scripts/
(coordinator, server, slow_queries, run_campaign, sandboxing, backends/) is
also internal to this directory and none of those use the underscore.
The _runtime.py name was inconsistent with the surrounding convention.

Generated-By: PostHog Code
Task-Id: 9ee4b87f-fce9-4d82-aded-510477a12e4a
